EPMLABSBRN-182 prepare resources for 2 series
move csv logic to csv package
