+++ src/main/kotlin/com/epam/brn/csv/CsvParser.kt
@@ -1,0 +1,9 @@
+package com.epam.brn.csv
+
+import com.fasterxml.jackson.databind.MappingIterator
+import java.io.InputStream
+
+interface CsvParser<Source> {
+
+    fun parseCsvFile(file: InputStream): MappingIterator<Source>
+}
