+++ src/main/kotlin/com/epam/brn/csv/UploadFromCsvStrategy.kt
@@ -1,0 +1,8 @@
+package com.epam.brn.csv
+
+import java.io.InputStream
+
+interface UploadFromCsvStrategy {
+
+    fun uploadFile(inputStream: InputStream): Map<String, String>
+}
