+++ src/main/kotlin/com/epam/brn/csv/converter/Converter.kt
@@ -1,0 +1,6 @@
+package com.epam.brn.csv.converter
+
+interface Converter<Source, Target> {
+
+    fun convert(source: Source): Target
+}
