+++ src/main/kotlin/com/epam/brn/csv/dto/GroupCsv.kt
@@ -1,0 +1,16 @@
+package com.epam.brn.csv.dto
+
+import com.fasterxml.jackson.annotation.JsonProperty
+import com.fasterxml.jackson.annotation.JsonPropertyOrder
+
+@JsonPropertyOrder(value = ["groupId", "name", "description"])
+data class GroupCsv(
+    @JsonProperty("groupId")
+    val groupId: Long,
+
+    @JsonProperty("name")
+    val name: String,
+
+    @JsonProperty("description")
+    val description: String
+)
