+++ src/main/kotlin/com/epam/brn/csv/firstSeries/commaSeparated/CommaSeparatedTaskCSVParserService.kt
@@ -1,0 +1,40 @@
+package com.epam.brn.csv.firstSeries.commaSeparated
+
+import com.epam.brn.csv.dto.TaskCsv
+import com.fasterxml.jackson.databind.MappingIterator
+import com.fasterxml.jackson.dataformat.csv.CsvMapper
+import com.fasterxml.jackson.dataformat.csv.CsvParser
+import java.io.InputStream
+import org.apache.commons.lang3.StringUtils
+import org.apache.logging.log4j.kotlin.logger
+import org.springframework.stereotype.Service
+
+@Service
+class CommaSeparatedTaskCSVParserService : com.epam.brn.csv.CsvParser<TaskCsv> {
+
+    val log = logger()
+
+    /**
+     * @param file - csv-file which should be convert to model
+     *
+     * @return csv-line to pair of model object and error. One pair values must be empty.
+     */
+
+    override fun parseCsvFile(file: InputStream): MappingIterator<TaskCsv> {
+        val csvMapper = CsvMapper().apply {
+            enable(CsvParser.Feature.TRIM_SPACES)
+        }
+
+        val csvSchema = csvMapper
+            .schemaFor(TaskCsv::class.java)
+            .withColumnSeparator(',')
+            .withLineSeparator(StringUtils.SPACE)
+            .withColumnReordering(true)
+            .withHeader()
+
+        return csvMapper
+            .readerWithTypedSchemaFor(TaskCsv::class.java)
+            .with(csvSchema)
+            .readValues<TaskCsv>(file)
+    }
+}
