--- src/main/kotlin/com/epam/brn/job/csv/task/UploadFromCsvStrategy.kt
@@ -1,8 +1,0 @@
-package com.epam.brn.job.csv.task
-
-import java.io.InputStream
-
-interface UploadFromCsvStrategy {
-
-    fun uploadFile(inputStream: InputStream): Map<String, String>
-}
