+++ src/main/kotlin/com/epam/brn/service/InitialDataLoader.kt
@@ -4,6 +4,17 @@ import com.epam.brn.constant.BrnRoles.AUTH_ROLE_ADMIN
 import com.epam.brn.constant.BrnRoles.AUTH_ROLE_USER
 import com.epam.brn.constant.ExerciseTypeEnum
 import com.epam.brn.constant.WordTypeEnum
+import com.epam.brn.csv.CsvMappingIteratorParser
+import com.epam.brn.csv.converter.impl.firstSeries.ExerciseCsvConverter
+import com.epam.brn.csv.converter.impl.firstSeries.GroupCsvConverter
+import com.epam.brn.csv.converter.impl.firstSeries.SeriesCsvConverter
+import com.epam.brn.csv.converter.impl.firstSeries.TaskCsv1SeriesConverter
+import com.epam.brn.csv.converter.impl.secondSeries.Exercise2SeriesConverter
+import com.epam.brn.csv.firstSeries.TaskCSVParser1SeriesService
+import com.epam.brn.csv.firstSeries.commaSeparated.CommaSeparatedExerciseCSVParserService
+import com.epam.brn.csv.firstSeries.commaSeparated.CommaSeparatedGroupCSVParserService
+import com.epam.brn.csv.firstSeries.commaSeparated.CommaSeparatedSeriesCSVParserService
+import com.epam.brn.csv.secondSeries.CSVParser2SeriesService
 import com.epam.brn.model.Authority
 import com.epam.brn.model.Exercise
 import com.epam.brn.model.ExerciseGroup
@@ -16,17 +27,6 @@ import com.epam.brn.repo.ExerciseRepository
 import com.epam.brn.repo.SeriesRepository
 import com.epam.brn.repo.TaskRepository
 import com.epam.brn.repo.UserAccountRepository
-import com.epam.brn.service.parsers.csv.CsvMappingIteratorParser
-import com.epam.brn.service.parsers.csv.converter.impl.firstSeries.ExerciseCsvConverter
-import com.epam.brn.service.parsers.csv.converter.impl.firstSeries.GroupCsvConverter
-import com.epam.brn.service.parsers.csv.converter.impl.firstSeries.SeriesCsvConverter
-import com.epam.brn.service.parsers.csv.converter.impl.firstSeries.TaskCsv1SeriesConverter
-import com.epam.brn.service.parsers.csv.converter.impl.secondSeries.Exercise2SeriesConverter
-import com.epam.brn.service.parsers.csv.firstSeries.TaskCSVParser1SeriesService
-import com.epam.brn.service.parsers.csv.firstSeries.commaSeparated.CommaSeparatedExerciseCSVParserService
-import com.epam.brn.service.parsers.csv.firstSeries.commaSeparated.CommaSeparatedGroupCSVParserService
-import com.epam.brn.service.parsers.csv.firstSeries.commaSeparated.CommaSeparatedSeriesCSVParserService
-import com.epam.brn.service.parsers.csv.secondSeries.CSVParser2SeriesService
 import java.io.InputStream
 import java.nio.file.Files
 import java.nio.file.Path
