--- src/main/kotlin/com/epam/brn/service/parsers/csv/CsvParser.kt
@@ -1,9 +1,0 @@
-package com.epam.brn.service.parsers.csv
-
-import com.fasterxml.jackson.databind.MappingIterator
-import java.io.InputStream
-
-interface CsvParser<Source> {
-
-    fun parseCsvFile(file: InputStream): MappingIterator<Source>
-}
