--- src/main/kotlin/com/epam/brn/service/parsers/csv/converter/Converter.kt
@@ -1,6 +1,0 @@
-package com.epam.brn.service.parsers.csv.converter
-
-interface Converter<Source, Target> {
-
-    fun convert(source: Source): Target
-}
