+++ src/main/resources/initFiles/1_series_full.csv
@@ -1,0 +1,64 @@
+level exerciseName orderNumber word audioFileName pictureFileName words wordType
+0 "Слова без шума" 1 бал no_noise/бал.mp3 pictures/бал.jpg (бум,быль,вить,гад,дуб) OBJECT
+0 "Слова без шума" 2 бум no_noise/бум.mp3 pictures/бум.jpg (бал,быль,вить,гад,дуб) OBJECT
+0 "Слова без шума" 3 быль no_noise/быль.mp3 pictures/быль.jpg (бал,бум,вить,гад,дуб) OBJECT
+0 "Слова без шума" 4 вить no_noise/вить.mp3 pictures/вить.jpg (бал,бум,быль,гад,дуб) OBJECT_ACTION
+0 "Слова без шума" 5 гад no_noise/гад.mp3 pictures/гад.jpg (бал,бум,быль,вить,дуб) OBJECT
+0 "Слова без шума" 6 дуб no_noise/дуб.mp3 pictures/дуб.jpg (бал,бум,быль,вить,гад) OBJECT
+1 "Однослоговые слова без шума" 1 бал no_noise/бал.mp3 pictures/бал.jpg (бам,сам,дам,зал,бум) OBJECT
+1 "Однослоговые слова без шума" 2 бум no_noise/бум.mp3 pictures/бум.jpg (зум,кум,шум,зуб,куб) OBJECT
+1 "Однослоговые слова без шума" 3 быль no_noise/быль.mp3 pictures/быль.jpg (пыль,соль,мыль,дыль,киль) OBJECT
+1 "Однослоговые слова без шума" 4 вить no_noise/вить.mp3 pictures/вить.jpg (бить,жить,мыль,выть,лить) OBJECT_ACTION
+1 "Однослоговые слова без шума" 5 гад no_noise/гад.mp3 pictures/гад.jpg (бад,сад,рад,бак,рать) OBJECT
+1 "Однослоговые слова без шума" 6 дуб no_noise/дуб.mp3 pictures/дуб.jpg (зуб,суд,мут,рут,пуд) OBJECT
+2 "Однослоговые слова без шума" 7 линь no_noise/линь.mp3 pictures/линь.jpg (пить,винь,лить,кинь,миль) OBJECT
+2 "Однослоговые слова без шума" 8 лис no_noise/лис.mp3 pictures/лис.jpg (пис,вис,мис,лиз,приз) OBJECT
+2 "Однослоговые слова без шума" 9 моль no_noise/моль.mp3 pictures/моль.jpg (боль,соль,воль,коль,доль) OBJECT
+2 "Однослоговые слова без шума" 10 пар no_noise/пар.mp3 pictures/пар.jpg (жар,кар,вар,мар,пат) OBJECT
+2 "Однослоговые слова без шума" 11 пять no_noise/пять.mp3 pictures/пять.jpg (жать,спать,врать,мать,пать) COUNT
+2 "Однослоговые слова без шума" 12 раб no_noise/раб.mp3 pictures/раб.jpg (жар,бар,сар,вар,тар) OBJECT
+3 "Однослоговые слова без шума" 13 рак no_noise/рак.mp3 pictures/рак.jpg (мар,жак,пак,нак,так) OBJECT
+3 "Однослоговые слова без шума" 14 рожь no_noise/рожь.mp3 pictures/рожь.jpg (ложь,мож,сож,кожа,вошь) OBJECT
+3 "Однослоговые слова без шума" 15 сеть no_noise/сеть.mp3 pictures/сеть.jpg (петь,жечь,речь,леть,веть) OBJECT
+3 "Однослоговые слова без шума" 16 топь no_noise/топь.mp3 pictures/топь.jpg (пот,сопь,ропь,моть,коть) OBJECT
+3 "Однослоговые слова без шума" 17 ход no_noise/ход.mp3 pictures/ход.jpg (сад,пот,рот,сот,кот) OBJECT
+3 "Однослоговые слова без шума" 18 шеф no_noise/шеф.mp3 pictures/шеф.jpg (пеф,сеф,рев,тев,пет) OBJECT
+3 "Однослоговые слова без шума" 19 мор no_noise/мор.mp3 pictures/мор.jpg (сор,лор,жор,пор,дор) OBJECT
+1 "Однослоговые слова малым шумом" 20 бал noise_6db/бал.mp3 pictures/бал.jpg (бам,сам,дам,зал,бак) OBJECT
+1 "Однослоговые слова малым шумом" 21 бум noise_6db/бум.mp3 pictures/бум.jpg (зум,кум,шум,зуб,куб) OBJECT
+1 "Однослоговые слова малым шумом" 22 быль noise_6db/быль.mp3 pictures/быль.jpg (пыль,соль,мыль,дыль,киль) OBJECT
+1 "Однослоговые слова малым шумом" 23 вить noise_6db/вить.mp3 pictures/вить.jpg (бить,жить,мыль,выть,лить) OBJECT_ACTION
+1 "Однослоговые слова малым шумом" 24 гад noise_6db/гад.mp3 pictures/гад.jpg (бад,сад,рад,бак,рать) OBJECT
+1 "Однослоговые слова малым шумом" 25 дуб noise_6db/дуб.mp3 pictures/дуб.jpg (зуб,суд,мут,рут,пуд) OBJECT
+2 "Однослоговые слова малым шумом" 26 линь noise_6db/линь.mp3 pictures/линь.jpg (пить,винь,лить,кинь,миль) OBJECT
+2 "Однослоговые слова малым шумом" 27 лис noise_6db/лис.mp3 pictures/лис.jpg (пис,вис,мис,лиз,приз) OBJECT
+2 "Однослоговые слова малым шумом" 28 моль noise_6db/моль.mp3 pictures/моль.jpg (боль,соль,воль,коль,доль) OBJECT
+2 "Однослоговые слова малым шумом" 29 пар noise_6db/пар.mp3 pictures/пар.jpg (жар,кар,вар,мар,пат) OBJECT
+2 "Однослоговые слова малым шумом" 30 пять noise_6db/пять.mp3 pictures/пять.jpg (жать,спать,врать,мать,пать) COUNT
+2 "Однослоговые слова малым шумом" 31 раб noise_6db/раб.mp3 pictures/раб.jpg (жар,бар,сар,вар,тар) OBJECT
+3 "Однослоговые слова малым шумом" 32 рак noise_6db/рак.mp3 pictures/рак.jpg (мар,жак,пак,нак,так) OBJECT
+3 "Однослоговые слова малым шумом" 33 рожь noise_6db/рожь.mp3 pictures/рожь.jpg (ложь,мож,сож,кожа,вошь) OBJECT
+3 "Однослоговые слова малым шумом" 34 сеть noise_6db/сеть.mp3 pictures/сеть.jpg (петь,жечь,речь,леть,веть) OBJECT
+3 "Однослоговые слова малым шумом" 35 топь noise_6db/топь.mp3 pictures/топь.jpg (пот,сопь,ропь,моть,коть) OBJECT
+3 "Однослоговые слова малым шумом" 36 ход noise_6db/ход.mp3 pictures/ход.jpg (сад,пот,рот,сот,кот) OBJECT
+3 "Однослоговые слова малым шумом" 37 шеф noise_6db/шеф.mp3 pictures/шеф.jpg (пеф,сеф,рев,тев,пет) OBJECT
+3 "Однослоговые слова малым шумом" 38 мор noise_6db/мор.mp3 pictures/мор.jpg (сор,лор,жор,пор,дор) OBJECT
+1 "Однослоговые слова c сильным шумом" 20 бал noise_9db/бал.mp3 pictures/бал.jpg (бам,сам,дам,зал,бак) OBJECT
+1 "Однослоговые слова c сильным шумом" 21 бум noise_9db/бум.mp3 pictures/бум.jpg (зум,кум,шум,зуб,куб) OBJECT
+1 "Однослоговые слова c сильным шумом" 22 быль noise_9db/быль.mp3 pictures/быль.jpg (пыль,соль,мыль,дыль,киль) OBJECT
+1 "Однослоговые слова c сильным шумом" 23 вить noise_9db/вить.mp3 pictures/вить.jpg (бить,жить,мыль,выть,лить) OBJECT_ACTION
+1 "Однослоговые слова c сильным шумом" 24 гад noise_9db/гад.mp3 pictures/гад.jpg (бад,сад,рад,бак,рать) OBJECT
+1 "Однослоговые слова c сильным шумом" 25 дуб noise_9db/дуб.mp3 pictures/дуб.jpg (зуб,суд,мут,рут,пуд) OBJECT
+2 "Однослоговые слова c сильным шумом" 26 линь noise_9db/линь.mp3 pictures/линь.jpg (пить,винь,лить,кинь,миль) OBJECT
+2 "Однослоговые слова c сильным шумом" 27 лис noise_9db/лис.mp3 pictures/лис.jpg (пис,вис,мис,лиз,приз) OBJECT
+2 "Однослоговые слова c сильным шумом" 28 моль noise_9db/моль.mp3 pictures/моль.jpg (боль,соль,воль,коль,доль) OBJECT
+2 "Однослоговые слова c сильным шумом" 29 пар noise_9db/пар.mp3 pictures/пар.jpg (жар,кар,вар,мар,пат) OBJECT
+2 "Однослоговые слова c сильным шумом" 30 пять noise_9db/пять.mp3 pictures/пять.jpg (жать,спать,врать,мать,пать) COUNT
+2 "Однослоговые слова c сильным шумом" 31 раб noise_9db/раб.mp3 pictures/раб.jpg (жар,бар,сар,вар,тар) OBJECT
+3 "Однослоговые слова c сильным шумом" 32 рак noise_9db/рак.mp3 pictures/рак.jpg (мар,жак,пак,нак,так) OBJECT
+3 "Однослоговые слова c сильным шумом" 33 рожь noise_9db/рожь.mp3 pictures/рожь.jpg (ложь,мож,сож,кожа,вошь) OBJECT
+3 "Однослоговые слова c сильным шумом" 34 сеть noise_9db/сеть.mp3 pictures/сеть.jpg (петь,жечь,речь,леть,веть) OBJECT
+3 "Однослоговые слова c сильным шумом" 35 топь noise_9db/топь.mp3 pictures/топь.jpg (пот,сопь,ропь,моть,коть) OBJECT
+3 "Однослоговые слова c сильным шумом" 36 ход noise_9db/ход.mp3 pictures/ход.jpg (сад,пот,рот,сот,кот) OBJECT
+3 "Однослоговые слова c сильным шумом" 37 шеф noise_9db/шеф.mp3 pictures/шеф.jpg (пеф,сеф,рев,тев,пет) OBJECT
+3 "Однослоговые слова c сильным шумом" 38 мор noise_9db/мор.mp3 pictures/мор.jpg (сор,лор,жор,пор,дор) OBJECT
